完善代理池，支持删除代理
